move StringFormatter to utils package
StringFormatter is a utility class that is not limited to supporting classes in list.
